perf(hutch,@packages/test-phase-runner): split prod and dev composition roots
app.ts imported ~40 `initInMemory*` factories from @packages/test-fixtures at
module top level and chose prod vs in-memory at runtime via PERSISTENCE, so the
test doubles were reachable from the bundled Lambda entry and esbuild could not
tree-shake them. Split the single composition root by environment:

- providers/prod-providers.ts (`initProdProviders`) wires only real providers;
  createHutchApp wires it unconditionally, so the lambda.main.ts import graph
  never reaches @packages/test-fixtures.
- providers/dev-providers.ts (`initDevProviders`) wires the in-memory providers
  for local dev and e2e and is the sole owner of the test-fixtures import edge.
- dev-app.ts (`createDevHutchApp`) reads PERSISTENCE and selects prod vs dev
  providers; server.main.ts uses it.

The shared page/middleware assembly (`assembleHutchApp`) is factored out and
called by both roots. Its provider parameter is typed
`HutchProviders = Omit<Parameters<typeof createApp>[0], AssemblyKeys>`, deriving
the contract from createApp's own parameter type so prod and dev cannot drift
and the looser dev fields (e.g. `stripePublishableKey: string | undefined`) are
absorbed rather than forcing the dev value to match a narrower prod-derived
type.

`dev:lambda` is removed: lambda.main.ts now composes prod providers
unconditionally, so routing it through the dev selector would put
@packages/test-fixtures back into the bundled entry's graph.

The `no-test-fixtures-in-runtime` guard's allowlist moves from `/app.ts` to
`/providers/dev-providers.ts` — the composition root that now legitimately
imports the fixtures — which also tightens it so a prod `app.ts` can no longer
import them.

The three new composition-root files carry the whole-file c8-ignore pragma
app.ts already used (no test-worthy logic); that pragma placement was approved
for this change.
