+++ projects/hutch/package.json
@@ -20,7 +20,6 @@
   "scripts": {
     "start": "NODE_ENV=production node dist/runtime/server.main.js",
     "dev": "npx --yes kill-port 35729 >/dev/null 2>&1; concurrently --kill-others \"livereload src --exts html,css,ts --wait 500\" \"node scripts/build-client-bundles.js --watch\" \"LIVERELOAD=true tsx watch src/runtime/server.main.ts\"",
-    "dev:lambda": "tsx watch src/runtime/lambda.main.ts",
     "permalink": "tsx src/runtime/web/pages/save/save-permalink.cli.main.ts",
     "backfill-registered-at": "tsx src/runtime/providers/auth/backfill-registered-at.cli.main.ts",
     "send-checkout-recovery-emails": "tsx src/runtime/checkout-recovery/send-checkout-recovery-emails.cli.main.ts",
